requirements.txt
flask
Flask-Restful
flask_sqlalchemy
psycopg2-binary